Move Toto-2 into the toto package and reuse the shared inference dataset
Colocate Toto2Model in toto/toto2.py alongside TotoModel (mirroring the
chronos/chronos2.py layout) and reuse TotoInferenceDataset instead of a
duplicate copy. The Toto-2 data loader now lives next to TotoDataLoader in
toto/dataloader.py.
