+++ timeseries/src/autogluon/timeseries/models/toto/__init__.py
@@ -1,3 +1,4 @@
 from .model import TotoModel
+from .toto2 import Toto2Model
 
-__all__ = ["TotoModel"]
+__all__ = ["Toto2Model", "TotoModel"]
