--- timeseries/src/autogluon/timeseries/models/toto2/__init__.py
@@ -1,3 +1,0 @@
-from .model import Toto2Model
-
-__all__ = ["Toto2Model"]
