+++ timeseries/tests/unittests/models/test_toto2.py
@@ -2,7 +2,7 @@ import numpy as np
 import pytest
 import torch
 
-from autogluon.timeseries.models.toto2.dataloader import Toto2DataLoader, Toto2InferenceDataset
+from autogluon.timeseries.models.toto.dataloader import Toto2DataLoader, TotoInferenceDataset
 
 from ..common import get_data_frame_with_item_index, get_data_frame_with_variable_lengths
 
@@ -53,7 +53,7 @@ class TestToto2Dataset:
     def test_when_dataset_iterated_then_context_has_correct_length(self, input_data_length, context_length):
         df = get_data_frame_with_item_index(["A", "B", "C", "D"], data_length=input_data_length)
 
-        dset = Toto2InferenceDataset(df, max_context_length=context_length)
+        dset = TotoInferenceDataset(df, max_context_length=context_length)
 
         for i in range(len(dset)):
             assert len(dset[i]) == min(context_length, input_data_length)
@@ -64,7 +64,7 @@ class TestToto2Dataset:
 
         df = get_data_frame_with_variable_lengths(item_id_to_length=item_id_to_length)
 
-        dset = Toto2InferenceDataset(df, max_context_length=max_data_length)
+        dset = TotoInferenceDataset(df, max_context_length=max_data_length)
 
         for i, item_length in zip(range(len(dset)), item_id_to_length.values()):
             assert len(dset[i]) == item_length
@@ -77,7 +77,7 @@ class TestToto2Dataloader:
             pytest.skip(reason="No GPU available")
 
         df = get_data_frame_with_item_index([f"item{x:03d}" for x in range(50)], data_length=100)
-        dataset = Toto2InferenceDataset(df, max_context_length=100)
+        dataset = TotoInferenceDataset(df, max_context_length=100)
         loader = Toto2DataLoader(dataset, batch_size=32, device=device)
 
         for batch in loader:
@@ -93,7 +93,7 @@ class TestToto2Dataloader:
         # 50 is not a multiple of 16 or 32
         data_length = 50
         df = get_data_frame_with_item_index(["A", "B", "C", "D"], data_length=data_length)
-        dataset = Toto2InferenceDataset(df, max_context_length=1000)
+        dataset = TotoInferenceDataset(df, max_context_length=1000)
         loader = Toto2DataLoader(dataset, batch_size=4, pad_to_multiple=pad_to_multiple, device=device)
 
         for batch in loader:
@@ -111,7 +111,7 @@ class TestToto2Dataloader:
         pad_to_multiple = 32
         data_length = 10  # shorter than a single patch
         df = get_data_frame_with_item_index(["A", "B"], data_length=data_length)
-        dataset = Toto2InferenceDataset(df, max_context_length=1000)
+        dataset = TotoInferenceDataset(df, max_context_length=1000)
         loader = Toto2DataLoader(dataset, batch_size=2, pad_to_multiple=pad_to_multiple, device=device)
 
         for batch in loader:
@@ -130,7 +130,7 @@ class TestToto2Dataloader:
         item_id_to_length = {"A": 1, "B": 10, "C": 10, "D": max_input_length}
         df = get_data_frame_with_variable_lengths(item_id_to_length=item_id_to_length)
 
-        dataset = Toto2InferenceDataset(df, max_context_length=1000)
+        dataset = TotoInferenceDataset(df, max_context_length=1000)
         loader = Toto2DataLoader(dataset, batch_size=4, device=device)
 
         for batch in loader:
@@ -146,7 +146,7 @@ class TestToto2Dataloader:
     def test_when_long_data_loaded_then_max_context_is_enforced(self, input_length, max_context_length):
         df = get_data_frame_with_item_index(["A", "B", "C", "D"], data_length=input_length)
 
-        dataset = Toto2InferenceDataset(df, max_context_length=max_context_length)
+        dataset = TotoInferenceDataset(df, max_context_length=max_context_length)
         loader = Toto2DataLoader(dataset, batch_size=4, device="cpu")
 
         for batch in loader:
@@ -159,7 +159,7 @@ class TestToto2Model:
     def test_predict_returns_correct_format(self, num_items, batch_size):
         from unittest.mock import patch
 
-        from autogluon.timeseries.models.toto2 import Toto2Model
+        from autogluon.timeseries.models.toto import Toto2Model
 
         item_index = [f"item{x:03d}" for x in range(num_items)]
         df = get_data_frame_with_item_index(item_index, data_length=50)
@@ -187,7 +187,7 @@ class TestToto2Model:
     def test_predict_interpolates_requested_quantile_levels(self):
         from unittest.mock import patch
 
-        from autogluon.timeseries.models.toto2 import Toto2Model
+        from autogluon.timeseries.models.toto import Toto2Model
 
         df = get_data_frame_with_item_index(["A", "B"], data_length=50)
